refactor: store 폴더 타입 리팩토링
