+++ src/components/atoms/Filter/index.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled'
 
 import { theme } from '@styles/theme'
 import { PropsType } from '@models/components/atoms/Filter'
-import { SelectedTagType } from '@models/favor'
+import { SelectedTagType } from '@models/common/favor'
 import { FilterIcon } from '@assets/icons/FilterIcon'
 import Svg from '../Svg'
