+++ src/components/atoms/GiftFilter/index.tsx
@@ -3,7 +3,7 @@ import { useRecoilState } from 'recoil'
 import styled from '@emotion/styled'
 
 import { theme } from '@styles/theme'
-import { SelectedPropsType, SelectedTagType } from '@models/favor'
+import { SelectedPropsType, SelectedTagType } from '@models/common/favor'
 import { PropsType } from '@models/components/atoms/GiftFilter'
 import { friendState } from '@libs/store/friend'
 import { FilterIcon } from '@assets/icons/FilterIcon'
