+++ src/components/profile/ProfileInfo/UserTagDataListItem.tsx
@@ -6,7 +6,7 @@ import { getTagAnswerData } from '@utils/getTagAnswerData'
 import { friendState } from '@libs/store/friend'
 import { question } from '@utils/question'
 import { TasteBoxVariantType } from '@models/apis/TasteType'
-import { SubCategoryNameType, SubCategoryValueType } from '@models/favor'
+import { SubCategoryNameType, SubCategoryValueType } from '@models/common/favor'
 import { ColorIndexVariantType } from '@models/components/atoms/Tag'
 import { UserTagDataPropsType } from '@models/components/Profile/profile'
 import { FlexBox } from '@components/layouts/FlexBox'
