+++ src/libs/store/question.ts
@@ -1,27 +1,26 @@
 import { atom } from 'recoil'
-import { DailyQuestionInfo } from '@models/apis/QuestionType'
-import { FavorAnswerRequest } from '@models/apis/TasteType'
-import { recoilPersist } from 'recoil-persist'
 
-const { persistAtom } = recoilPersist()
-const initialState: DailyQuestionInfo = {
+import { DailyQuestionInfoType } from '@models/apis/QuestionType'
+import { FavorAnswerRequestType } from '@models/apis/TasteType'
+
+const initialState: DailyQuestionInfoType = {
   questionId: 0,
   content: '',
   category: 'FOOD',
   loadingData: '',
 }
 
-export const initialAnswerState: FavorAnswerRequest = {
+export const initialAnswerState: FavorAnswerRequestType = {
   categoryName: '',
   favorAnswerDtos: [],
 }
 
-export const questionState = atom<DailyQuestionInfo>({
+export const questionState = atom<DailyQuestionInfoType>({
   key: 'dailyQuestion',
   default: initialState,
 })
 
-export const answerState = atom<FavorAnswerRequest>({
+export const answerState = atom<FavorAnswerRequestType>({
   key: 'favorAnswer',
   default: initialAnswerState,
 })
