+++ src/models/common/favor.ts
