+++ src/models/common/favorPriority.ts
@@ -1,0 +1,6 @@
+import { TasteType } from '@models/apis/TasteType'
+
+export type FavorPriorityType = {
+  taste: TasteType
+  priority: number
+}
