+++ src/models/stores/auth.ts
@@ -1,0 +1,9 @@
+import { UserInfoTokenType } from '@models/apis/UserType'
+
+export type AuthStateType = {
+  isAuthenticated: boolean
+  loginType: 'APPLE' | 'KAKAO'
+  callbackUrl: string
+  accessToken: string
+  userProfile: UserInfoTokenType
+}
