+++ src/models/stores/date.ts
@@ -1,0 +1,7 @@
+import { TodayKeyType } from '@models/components/atoms/DayWeek'
+
+export type WeeklyDateType = {
+  today: TodayKeyType
+  selectedDate: TodayKeyType
+  dateString: string
+}
