+++ src/models/stores/present.ts
@@ -1,0 +1,9 @@
+export type FilterType = {
+  filter: string
+  filterValue: string
+}
+
+export type PriceType = {
+  price: string
+  priceValue: string
+}
