+++ src/models/stores/profile.ts
@@ -1,0 +1,9 @@
+import { TodayCategoryValueType } from '@models/components/atoms/TodayCategory'
+
+export type ProfileStateType = {
+  value: boolean
+  isMenuOpen: boolean
+  isEditImageMenuOpen: boolean
+  isEdit: boolean
+  pastTodayCategory: TodayCategoryValueType
+}
